My robot
The robot works and it beats the other bots most of the times.
It iterates through all moves within our vision range (has no memory yet) and chooses the 2-step-path with the lowest distance to the goldpot which is valid. 
The distance is calculated with the sqrt(dist(x)**2 + dist(y)**2).
Depending on the map the robot is stuck a lot; however, if a quite direct path is possible it usually outpaces the other bots.