Final Robot for the meeting
The Robot saves the map as a dict in a text file 
It calculates the shortest path with networkx

Several constraints are used to optimize the number of moves: maximum 10, but always only until an unknown tile would be used. 
Furthermore, to avoid being stuck at ~0 gold the cost for the moves cannot exceed 10% of our total gold, except if the gold pot can be reached within 5 steps. 
The Robot does not move if it has less than 15 gold and the gold pot is more than half the map away.